+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 # Version 0.5.6
 
+* Issue #12289 - Added DO driver to generate stable dissolved oxygen stream from CTD with attached DO
 * Issue #12323 - Added support for UI generated ZPLSC C-Series echograms.
 
 # Version 0.5.5